fix infinite timer ui and logic
