+++ components/bsb/timer/background/api/src/commonMain/kotlin/com/flipperdevices/bsb/timer/background/statefactory/model/IterationType.kt
@@ -1,0 +1,11 @@
+package com.flipperdevices.bsb.timer.background.statefactory.model
+
+enum class IterationType {
+    WORK,
+    REST,
+    LONG_REST,
+    WAIT_AFTER_WORK,
+    WAIT_AFTER_REST;
+
+    companion object
+}
