--- components/bsb/timer/background/api/src/commonMain/kotlin/com/flipperdevices/bsb/timer/background/util/ControlledTimerStateFactory.kt
@@ -1,269 +1,0 @@
-package com.flipperdevices.bsb.timer.background.util
-
-import com.flipperdevices.bsb.dao.model.TimerDuration
-import com.flipperdevices.bsb.dao.model.TimerSettings
-import com.flipperdevices.bsb.timer.background.model.ControlledTimerState
-import com.flipperdevices.bsb.timer.background.model.TimerTimestamp
-import com.flipperdevices.core.log.TaggedLogger
-import com.flipperdevices.core.log.error
-import com.flipperdevices.core.log.wtf
-import kotlinx.datetime.Clock
-import kotlin.time.Duration
-import kotlin.time.Duration.Companion.seconds
-
-private val logger = TaggedLogger("ControlledTimerStateFactory")
-
-enum class IterationType {
-    WORK, REST, LONG_REST, WAIT_AFTER_WORK, WAIT_AFTER_REST
-}
-
-sealed interface IterationData {
-    val iterationType: IterationType
-    val startOffset: Duration
-    val duration: Duration
-
-    data object Infinite : IterationData {
-        override val startOffset = 0.seconds
-        override val iterationType = IterationType.WORK
-        override val duration = Duration.INFINITE
-    }
-
-    data class Pending(
-        override val startOffset: Duration,
-        override val iterationType: IterationType,
-        override val duration: Duration
-    ) : IterationData
-
-    data class Default(
-        override val startOffset: Duration,
-        override val duration: Duration,
-        override val iterationType: IterationType
-    ) : IterationData
-}
-
-@Suppress("MagicNumber")
-private fun getIterationTypeByIndex(i: Int): IterationType {
-    return when {
-        i % 2 == 0 -> IterationType.WORK
-        i % 3 == 2 -> IterationType.LONG_REST
-        i % 2 == 1 -> IterationType.REST
-        else -> {
-            logger.error { "#buildIterationList could not calculate next IterationType for index $i" }
-            IterationType.WORK
-        }
-    }
-}
-
-@Suppress("LongMethod", "CyclomaticComplexMethod")
-fun TimerSettings.buildIterationList(): List<IterationData> {
-    val totalTime = when (val localTotalTime = totalTime) {
-        is TimerDuration.Finite -> localTotalTime.instance
-        TimerDuration.Infinite -> {
-            return listOf(IterationData.Infinite)
-        }
-    }
-    if (!intervalsSettings.isEnabled) {
-        return listOf(
-            IterationData.Default(
-                startOffset = 0.seconds,
-                duration = totalTime,
-                iterationType = IterationType.WORK
-            )
-        )
-    }
-    val list = buildList<IterationData> {
-        var timeLeft = totalTime
-        var i = 0
-        while (timeLeft > 0.seconds) {
-            val type = getIterationTypeByIndex(i)
-            val iterationTypeDuration = when (type) {
-                IterationType.WORK -> intervalsSettings.work
-                IterationType.REST -> intervalsSettings.rest
-                IterationType.LONG_REST -> intervalsSettings.longRest
-
-                IterationType.WAIT_AFTER_REST,
-                IterationType.WAIT_AFTER_WORK -> {
-                    logger.wtf { "#buildIterationList wait iteration appeared inside getIterationTypeByIndex" }
-                    0.seconds
-                }
-            }.coerceAtMost(timeLeft)
-
-            val isNoTimeForWorkLeft = timeLeft <= iterationTypeDuration &&
-                type == IterationType.WORK
-
-            val isNoTimeForShortRestLeft =
-                timeLeft <= (iterationTypeDuration + intervalsSettings.work) &&
-                    type == IterationType.REST
-
-            val isLongRestNeedMoreTimeThanTimeLeft =
-                timeLeft <= (iterationTypeDuration + intervalsSettings.longRest) &&
-                    type == IterationType.LONG_REST
-
-            val resolvedIterationType = when {
-                isNoTimeForWorkLeft
-                    .or(isNoTimeForShortRestLeft)
-                    .or(isLongRestNeedMoreTimeThanTimeLeft) -> IterationType.LONG_REST
-
-                else -> type
-            }
-
-            if (resolvedIterationType == IterationType.LONG_REST &&
-                lastOrNull()?.iterationType == IterationType.LONG_REST
-            ) {
-                timeLeft -= iterationTypeDuration
-                i += 1
-                continue
-            }
-
-            IterationData.Default(
-                startOffset = totalTime - timeLeft,
-                iterationType = resolvedIterationType,
-                duration = when {
-                    isNoTimeForShortRestLeft ->
-                        iterationTypeDuration
-                            .plus(intervalsSettings.work)
-                            .coerceAtMost(timeLeft)
-                            .also { timeLeft -= intervalsSettings.work }
-
-                    isLongRestNeedMoreTimeThanTimeLeft ->
-                        iterationTypeDuration
-                            .plus(intervalsSettings.longRest)
-                            .coerceAtMost(timeLeft)
-                            .also { timeLeft -= intervalsSettings.work }
-
-                    isNoTimeForWorkLeft -> iterationTypeDuration
-                    else -> iterationTypeDuration
-                }
-            ).run(::add)
-            if (listOf(
-                    isNoTimeForWorkLeft,
-                    isNoTimeForShortRestLeft,
-                    isLongRestNeedMoreTimeThanTimeLeft
-                ).all { !it }
-            ) {
-                if (type == IterationType.WORK && !intervalsSettings.autoStartRest) {
-                    IterationData.Pending(
-                        startOffset = totalTime - timeLeft + iterationTypeDuration,
-                        iterationType = IterationType.WAIT_AFTER_WORK,
-                        duration = Duration.INFINITE
-                    ).run(::add)
-                }
-                if (type in listOf(
-                        IterationType.REST,
-                        IterationType.LONG_REST
-                    ) && !intervalsSettings.autoStartWork
-                ) {
-                    IterationData.Pending(
-                        startOffset = totalTime - timeLeft + iterationTypeDuration,
-                        iterationType = IterationType.WAIT_AFTER_REST,
-                        duration = Duration.INFINITE
-                    ).run(::add)
-                }
-            }
-
-            timeLeft -= iterationTypeDuration
-            i += 1
-        }
-    }.toMutableList()
-    if (list.isEmpty()) {
-        logger.wtf { "#buildIterationList was empty for $this" }
-        return list
-    }
-    return list
-}
-
-private val TimerSettings.maxIterationCount: Int
-    get() = buildIterationList()
-        .count { data -> data.iterationType == IterationType.WORK }
-
-@Suppress("LongMethod")
-fun TimerTimestamp.toState(): ControlledTimerState {
-    if (this !is TimerTimestamp.Running) {
-        return ControlledTimerState.NotStarted
-    }
-    val iterationList = settings.buildIterationList()
-    val now = pause ?: Clock.System.now()
-
-    // Filter only data which is not yet started
-    val iterationsDataLeft = iterationList
-        .filter { data ->
-            when (data) {
-                is IterationData.Default -> {
-                    now <= start.plus(data.startOffset).plus(data.duration)
-                }
-
-                is IterationData.Pending -> {
-                    confirmNextStepClick < start.plus(data.startOffset)
-                }
-
-                is IterationData.Infinite -> {
-                    return ControlledTimerState.InProgress.Running.Work(
-                        timeLeft = TimerDuration.Infinite,
-                        isOnPause = pause != null,
-                        timerSettings = settings,
-                        currentIteration = 1,
-                        maxIterations = 1
-                    )
-                }
-            }
-        }
-    val currentIterationData = iterationsDataLeft.firstOrNull()
-
-    if (currentIterationData == null) {
-        return ControlledTimerState.Finished(timerSettings = settings)
-    }
-
-    val iterationCountLeft = settings.maxIterationCount
-        .minus(
-            iterationsDataLeft
-                .filterIndexed { i, _ -> i != 0 }
-                .count { data -> data.iterationType == IterationType.WORK }
-        )
-
-    val currentIterationTypeTimeLeft = start
-        .plus(currentIterationData.startOffset)
-        .plus(currentIterationData.duration)
-        .minus(now)
-
-    return when (currentIterationData.iterationType) {
-        IterationType.WORK -> ControlledTimerState.InProgress.Running.Work(
-            timeLeft = TimerDuration(currentIterationTypeTimeLeft),
-            isOnPause = pause != null,
-            timerSettings = settings,
-            currentIteration = iterationCountLeft,
-            maxIterations = settings.maxIterationCount,
-        )
-
-        IterationType.REST -> ControlledTimerState.InProgress.Running.Rest(
-            timeLeft = TimerDuration(currentIterationTypeTimeLeft),
-            isOnPause = pause != null,
-            timerSettings = settings,
-            currentIteration = iterationCountLeft,
-            maxIterations = settings.maxIterationCount,
-        )
-
-        IterationType.LONG_REST -> ControlledTimerState.InProgress.Running.LongRest(
-            timeLeft = TimerDuration(currentIterationTypeTimeLeft),
-            isOnPause = pause != null,
-            timerSettings = settings,
-            currentIteration = iterationCountLeft,
-            maxIterations = settings.maxIterationCount,
-        )
-
-        IterationType.WAIT_AFTER_REST -> ControlledTimerState.InProgress.Await(
-            timerSettings = settings,
-            currentIteration = iterationCountLeft,
-            maxIterations = settings.maxIterationCount,
-            pausedAt = start.plus(currentIterationData.startOffset),
-            type = ControlledTimerState.InProgress.AwaitType.AFTER_REST
-        )
-
-        IterationType.WAIT_AFTER_WORK -> ControlledTimerState.InProgress.Await(
-            timerSettings = settings,
-            currentIteration = iterationCountLeft,
-            maxIterations = settings.maxIterationCount,
-            pausedAt = start.plus(currentIterationData.startOffset),
-            type = ControlledTimerState.InProgress.AwaitType.AFTER_WORK
-        )
-    }
-}
