+++ components/bsb/timer/background/impl/src/commonTest/kotlin/com/flipperdevices/bsb/timer/background/api/util/ControlledTimerStateFactoryTest.kt
@@ -2,9 +2,9 @@ package com.flipperdevices.bsb.timer.background.api.util
 
 import com.flipperdevices.bsb.dao.model.TimerSettings
 import com.flipperdevices.bsb.dao.model.TimerSettingsId
-import com.flipperdevices.bsb.timer.background.util.IterationData
-import com.flipperdevices.bsb.timer.background.util.IterationType
-import com.flipperdevices.bsb.timer.background.util.buildIterationList
+import com.flipperdevices.bsb.timer.background.statefactory.buildIterationList
+import com.flipperdevices.bsb.timer.background.statefactory.model.IterationData
+import com.flipperdevices.bsb.timer.background.statefactory.model.IterationType
 import kotlin.test.Test
 import kotlin.test.assertContentEquals
 import kotlin.time.Duration.Companion.hours
